relocate testcase parser
